Quitado un último error.

+++ trunk/Proyecto_ProgramacionIV/src/bomberman/ventanas/VentanaCargar.java
@@ -165,18 +165,12 @@ public class VentanaCargar extends JDialog implements ActionListener {
 	 */
 	public void setVisible(boolean estado) {
 		super.setVisible(estado);
-		/*
-		 * Lo usamos en el EDT para no crear problemas.
-		 */
-			java.awt.EventQueue.invokeLater(new Runnable() {
-			    @Override
-			    public void run() {
+		if(estado){
 			    	tmModel.deleteAllRows();
 					ArrayList<PuntuGeneral> partGuardas = AccesoPuntuGen
 							.getPartidasGuardadas();
 					for (PuntuGeneral tempPunt : partGuardas)
 						tmModel.añadirFila(tempPunt);
 			    }
-			});
 	}
 }
